# Perf optimization for SQL GROUP BY ORDER BY