--- a/pinot-common/src/main/java/org/apache/pinot/common/metrics/BrokerMeter.java
+++ b/pinot-common/src/main/java/org/apache/pinot/common/metrics/BrokerMeter.java
@@ -65,7 +65,7 @@ public enum BrokerMeter implements AbstractMetrics.Meter {
   ENTRIES_SCANNED_POST_FILTER("documents", false),
 
   NUM_RESIZES("resizes", false),
-  RESIZE_TIME_MS("documents", false),
+  RESIZE_TIME_MS("resizeTimeMs", false),
 
   REQUEST_CONNECTION_TIMEOUTS("timeouts", false),
   HELIX_ZOOKEEPER_RECONNECTS("reconnects", true),