--- a/pinot-core/src/main/java/org/apache/pinot/core/data/table/SimpleIndexedTable.java
+++ b/pinot-core/src/main/java/org/apache/pinot/core/data/table/SimpleIndexedTable.java
@@ -37,16 +37,15 @@ import org.slf4j.LoggerFactory;
 public class SimpleIndexedTable extends IndexedTable {
   private static final Logger LOGGER = LoggerFactory.getLogger(SimpleIndexedTable.class);
 
-  private final Map<Key, Record> _lookupMap;
+  private Map<Key, Record> _lookupMap;
   private Iterator<Record> _iterator;
 
   private boolean _noMoreNewRecords = false;
   private int _numResizes = 0;
-  private long _resizeTime = 0;
+  private long _resizeTimeMs = 0;
 
-  public SimpleIndexedTable(DataSchema dataSchema, QueryContext queryContext, int capacity) {
-    super(dataSchema, queryContext, capacity);
-
+  public SimpleIndexedTable(DataSchema dataSchema, QueryContext queryContext, int trimSize, int trimThreshold) {
+    super(dataSchema, queryContext, trimSize, trimThreshold);
     _lookupMap = new HashMap<>();
   }
 
@@ -56,7 +55,6 @@ public class SimpleIndexedTable extends IndexedTable {
   @Override
   public boolean upsert(Key key, Record newRecord) {
     Preconditions.checkNotNull(key, "Cannot upsert record with null keys");
-
     if (_noMoreNewRecords) { // allow only existing record updates
       _lookupMap.computeIfPresent(key, (k, v) -> {
         Object[] existingValues = v.getValues();
@@ -83,10 +81,10 @@ public class SimpleIndexedTable extends IndexedTable {
         }
       });
 
-      if (_lookupMap.size() >= _maxCapacity) {
+      if (_lookupMap.size() >= _trimThreshold) {
         if (_hasOrderBy) {
           // reached max capacity, resize
-          resize(_capacity);
+          resize(_trimSize);
         } else {
           // reached max capacity and no order by. No more new records will be accepted
           _noMoreNewRecords = true;
@@ -97,36 +95,27 @@ public class SimpleIndexedTable extends IndexedTable {
   }
 
   private void resize(int trimToSize) {
-
     long startTime = System.currentTimeMillis();
-
-    _tableResizer.resizeRecordsMap(_lookupMap, trimToSize);
-
+    _lookupMap = _tableResizer.resizeRecordsMap(_lookupMap, trimToSize);
     long endTime = System.currentTimeMillis();
     long timeElapsed = endTime - startTime;
-
     _numResizes++;
-    _resizeTime += timeElapsed;
+    _resizeTimeMs += timeElapsed;
   }
 
   private List<Record> resizeAndSort(int trimToSize) {
-
     long startTime = System.currentTimeMillis();
-
-    List<Record> sortedRecords = _tableResizer.resizeAndSortRecordsMap(_lookupMap, trimToSize);
-
+    List<Record> sortedRecords = _tableResizer.sortRecordsMap(_lookupMap, trimToSize);
     long endTime = System.currentTimeMillis();
     long timeElapsed = endTime - startTime;
-
     _numResizes++;
-    _resizeTime += timeElapsed;
-
+    _resizeTimeMs += timeElapsed;
     return sortedRecords;
   }
 
   @Override
   public int size() {
-    return _lookupMap.size();
+    return _sortedRecords == null ? _lookupMap.size() : _sortedRecords.size();
   }
 
   @Override
@@ -136,22 +125,28 @@ public class SimpleIndexedTable extends IndexedTable {
 
   @Override
   public void finish(boolean sort) {
-
     if (_hasOrderBy) {
-
       if (sort) {
-        List<Record> sortedRecords = resizeAndSort(_capacity);
-        _iterator = sortedRecords.iterator();
+        _sortedRecords = resizeAndSort(_trimSize);
+        _iterator = _sortedRecords.iterator();
       } else {
-        resize(_capacity);
+        resize(_trimSize);
       }
-      LOGGER
-          .debug("Num resizes : {}, Total time spent in resizing : {}, Avg resize time : {}", _numResizes, _resizeTime,
-              _numResizes == 0 ? 0 : _resizeTime / _numResizes);
+      LOGGER.debug("Num resizes : {}, Total time spent in resizing : {}, Avg resize time : {}, trimSize: {}, trimThreshold: {}",
+          _numResizes, _resizeTimeMs, _numResizes == 0 ? 0 : _resizeTimeMs / _numResizes, _trimSize, _trimThreshold);
     }
-
     if (_iterator == null) {
       _iterator = _lookupMap.values().iterator();
     }
+  }
+
+  @Override
+  public int getNumResizes() {
+    return _numResizes;
+  }
+
+  @Override
+  public long getResizeTimeMs() {
+    return _resizeTimeMs;
   }
 }