… pinot-core/src/main/java/org/apache/pinot/core/operator/blocks/IntermediateResultsBlock.java
… pinot-core/src/main/java/org/apache/pinot/core/operator/combine/GroupByOrderByCombineOperator.java
… pinot-core/src/main/java/org/apache/pinot/core/plan/maker/InstancePlanMakerImplV2.java
… pinot-core/src/main/java/org/apache/pinot/core/query/reduce/BrokerReduceService.java
… pinot-core/src/main/java/org/apache/pinot/core/query/reduce/DataTableReducerContext.java
… pinot-core/src/main/java/org/apache/pinot/core/query/reduce/GroupByDataTableReducer.java
… pinot-core/src/main/java/org/apache/pinot/core/query/scheduler/QueryScheduler.java
… pinot-core/src/test/java/org/apache/pinot/core/data/table/IndexedTableTest.java
… pinot-core/src/test/java/org/apache/pinot/core/data/table/TableResizerTest.java
… pinot-core/src/test/java/org/apache/pinot/core/operator/combine/CombineSlowOperatorsTest.java
… pinot-core/src/test/java/org/apache/pinot/core/operator/combine/SelectionCombineOperatorTest.java
… pinot-core/src/test/java/org/apache/pinot/core/plan/CombinePlanNodeTest.java
… pinot-perf/src/main/java/org/apache/pinot/perf/BenchmarkCombineGroupBy.java
… pinot-perf/src/main/java/org/apache/pinot/perf/BenchmarkIndexedTable.java