This PR optimizes the performance of SQL GROUP BY ORDER BY.

**Problem:**

We recently observed that GROUP BY ORDER BY queries having a large (we are having as high as 166million) numDocsScanned are 5x-6x slower as compared to PQL. Removing ORDER BY from the queries made the performance on par or even faster than PQL in some cases. So ORDER BY code was the reason for degradation during server-combine and broker-reduce phases since segment level execution between PQL and SQL is the same for GROUP BY ORDER BY.

Couple of factors in the order by code are contributing to degradation:

- More resizes: Resizes are essentially not the internal resize of the ConcurrentHashMap used by IndexedTable. The resize here is the trimming operation done by TableResizer everytime the map reaches the trimThreshold. The resize requires inserting items in PQ (sorting) and evicting the `trimThreshold - trimSize` number of items from the map. 

- retainAll() - The TableResizer's resizeRecordsMap() method is invoked under 2 circumstances (1) resize during upsert, (2) resize during finish(false) from server combine. If TableResizer builds a PQ to retain the records, it then invokes retainAll() method on hashmap which we have seen to take > 1sec

- readLock.lock and readLock.unlock - The upsert method takes a readlock to prevent the race condition that might arise due to another thread concurrently trimming (under write lock). The overhead of this operation is non-negligible for very high number of records. The total cumulative time spent in lock and unlock operations was just under 1sec
`

  | PQL | SQL | SQL optimized
-- | -- | -- | --
**Server Combine**
trimSize for server combine | max(TOP * 5, 5000) | max(LIMIT * 5, 5000) | max(LIMIT * 5, 5000)
trimThreshold for segment combine | trimSize * 4 | trimSize <= 100_000 trimThreshold = 1m else trimThreshold = trimSize * 1.2 | trimThreshold is configurable
Num resizes/sort during segment execution | 0 | 0 | 0
Num resizes/sort during server combine | At most once - when the server combine is done, if numGroups > trimThreshold, trim to trimSize. | During upsert every time we hit trimThreshold, trim to trimSize.During finish, we trim again to trimSize | During upsert, no trimming. During finish, if numGroups > trimThreshold, trim to trimSize. Finish is called when the server combine is over.
**Broker Reduce**
trimSize for broker reduce | TOP | max(LIMIT * 5, 5000) | LIMIT
trimThreshold for broker reduce | N.A | trimSize <= 100_000 trimThreshold = 1m else trimThreshold = trimSize * 1.2 | N.A
Num resizes/sort during broker combine | Exactly once - after adding results from all DataTables to the finalMap, trim to TOP | We upsert rows from each DataTable into IndexedTable. During upsert every time we hit trimThreshold, trim to trimSize. During finish, we resize again to trimSize Finally use LIMIT number of rows from the trimmed set as the final result set | During upsert, no trimming. Exactly once - when finish is called after upserting rows from each DataTable into indexed table. Trimmed to LIMIT

`

For a given SQL query with LIMIT 50000 and numDocsScanned 160million:

- trimSize -> max (50000 * 5, 5000) -> 250000
- trimThreshold -> trimSize * 1.2 -> 300000
- num resizes(trims) during server combine - ~20
- cumulative resize/trim time during server combine - 3000ms
- num resizes (trims) during broker reduce - ~10
- cumulative resize/trim time during broker reduce - 1600ms
- readLock.lock, unlock - 900ms
- retainAll -  1100ms

Total SQL query time ~7000ms
Total PQL query time ~1800ms

**Solution**

The solution is aimed at bringing the resize/trim code on par with PQL to get similar performance for such use cases

- To better control the frequency of trimming (and accuracy, latency), introduce a new server side and broker side config for trimThreshold. To be used for SQL group by. 

- trimThreshold = configuredTrimThreshold

- trimSize = limit N * 5 (same as PQL, not changed by this PR)

- trimSize = min (trimSize, trimThreshold / 2) - to protect the server with queries having high LIMIT N

- If the configured value is infinite (1B), we use a sub-class of ConcurrentIndexedTable that doesn't use any locks in upsert() method since we know that trimThreshold being infinite wouldn't require any trimming in upsert. 

- Instead of retainAll(), build a new result map with the records to be retained just like it is done in PQL. 

- During server combine, trim at most once when indexedTable.finish (false) is called by combiner if numGroups > trimThreshold. Trim to trimSize

- During broker reduce, trim exactly once when indexedTable.finish(true) is called. Trim to LIMIT N

For the same query, with the above changes

SQL - 2200ms (multiple runs also give < 2secs)
PQL - 1800ms