@mayankshriv 2020-11-03
> If there is no resize during the upsert, how do you prevent server from running out of memory?
> One big difference between PQL and SQL is that for PQL we limit the groups in the map (trimming during upsert is desired, but not possible for PQL semantic). For SQL we want to trim the result during upsert to have the accurate result and also prevent OOM. We can increase the trimming threshold to reduce the number of trimmings.

+1, eliminating resizing during upsert is dangerous and can cause OOM. PQL has a config to bypass the truncation that certain use cases are using, perhaps the same can be used for reducing num resize during upserts (defaulting to a tuned value that does not require a general case to explicitly set a value).