@siddharthteotia 2020-11-17
> Could you confirm that this PR has behavior change only for large group-bys? For small/medium group-bys is it identical behavior? If not, could you call out the changes?

- For segment level, there is no change (even the PR doesn't introduce any change at segment level)
- For server and broker level, now the SQL also uses the 2 * 100k (configurable) as the trimThreshold. If any current use case is under 100k, there will not be any change.

Also, with this change, for SQL GROUP BY we can now use the configuration and fine-tune on a per use case basis to decrease the frequency of trimming (during upsert) just like it can be reduced to 0 for use cases with large group by.